cleaning up evals.env logic in extension.ts (#3045)
